+++ src/body/Stash Batch Query Edit.user.js
@@ -17,10 +17,12 @@
 
     let running = false;
     const buttons = [];
+    let maxCount = 0;
 
     function run(videoExtensions) {
         if (!running) return;
         const button = buttons.pop();
+        stash.setProgress((maxCount - buttons.length) / maxCount * 100);
         if (button) {
             const searchItem = getClosestAncestor(button, '.search-item');
             const {
@@ -97,12 +99,14 @@
         btn.classList.remove('btn-primary');
         btn.classList.add('btn-danger');
         running = true;
+        stash.setProgress(0);
         buttons.length = 0;
         for (const button of document.querySelectorAll('.btn.btn-primary')) {
             if (button.innerText === 'Search') {
                 buttons.push(button);
             }
         }
+        maxCount = buttons.length;
         const reqData = {
             "variables": {},
             "query": `query Configuration {
@@ -123,6 +127,7 @@
         btn.classList.remove('btn-danger');
         btn.classList.add('btn-primary');
         running = false;
+        stash.setProgress(0);
     }
 
     stash.addEventListener('tagger:mutations:header', evt => {