+++ src/body/Stash Batch Save.user.js
@@ -14,11 +14,13 @@
 
     let running = false;
     const buttons = [];
+    let maxCount = 0;
     let sceneId = null;
 
     function run() {
         if (!running) return;
         const button = buttons.pop();
+        stash.setProgress((maxCount - buttons.length) / maxCount * 100);
         if (button) {
             const searchItem = getClosestAncestor(button, '.search-item');
             const { id } = stash.parseSearchItem(searchItem);
@@ -63,12 +65,14 @@
         btn.classList.remove('btn-primary');
         btn.classList.add('btn-danger');
         running = true;
+        stash.setProgress(0);
         buttons.length = 0;
         for (const button of document.querySelectorAll('.btn.btn-primary')) {
             if (button.innerText === 'Save') {
                 buttons.push(button);
             }
         }
+        maxCount = buttons.length;
         stash.addEventListener('stash:response', processSceneUpdate);
         run();
     }
@@ -78,6 +82,7 @@
         btn.classList.remove('btn-danger');
         btn.classList.add('btn-primary');
         running = false;
+        stash.setProgress(0);
         sceneId = null;
         stash.removeEventListener('stash:response', processSceneUpdate);
     }