+++ src/body/Stash Batch Search.user.js
@@ -17,10 +17,12 @@
 
     let running = false;
     const buttons = [];
+    let maxCount = 0;
 
     function run() {
         if (!running) return;
         const button = buttons.pop();
+        stash.setProgress((maxCount - buttons.length) / maxCount * 100);
         if (button) {
             if (!button.disabled) {
                 button.click();
@@ -56,12 +58,14 @@
         btn.classList.remove('btn-primary');
         btn.classList.add('btn-danger');
         running = true;
+        stash.setProgress(0);
         buttons.length = 0;
         for (const button of document.querySelectorAll('.btn.btn-primary')) {
             if (button.innerText === 'Search') {
                 buttons.push(button);
             }
         }
+        maxCount = buttons.length;
         run();
     }
 
@@ -70,6 +74,7 @@
         btn.classList.remove('btn-danger');
         btn.classList.add('btn-primary');
         running = false;
+        stash.setProgress(0);
     }
 
     stash.addEventListener('page:performers', function () {